fix : 룸 삭제 API 수정 - 기존 : 룸 엔티티만 삭제 - 개선 : 삭제 시 연관된 이미지 파일 삭제, 연관된 아이템 있을 시 삭제 불가

+++ src/main/java/com/hoo/aoo/admin/application/service/room/DeleteRoomService.java
@@ -1,8 +1,14 @@
 package com.hoo.aoo.admin.application.service.room;
 
 import com.hoo.aoo.admin.application.port.in.room.DeleteRoomUseCase;
+import com.hoo.aoo.admin.application.port.out.item.FindItemPort;
 import com.hoo.aoo.admin.application.port.out.room.DeleteRoomPort;
+import com.hoo.aoo.admin.application.port.out.room.FindRoomPort;
+import com.hoo.aoo.admin.application.service.AdminErrorCode;
+import com.hoo.aoo.admin.application.service.AdminException;
+import com.hoo.aoo.admin.domain.room.Room;
 import com.hoo.aoo.common.application.port.in.MessageDto;
+import com.hoo.aoo.file.application.port.in.DeleteFileUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,12 +17,23 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class DeleteRoomService implements DeleteRoomUseCase {
 
+    private final FindRoomPort findRoomPort;
+    private final FindItemPort findItemPort;
     private final DeleteRoomPort deleteRoomPort;
+    private final DeleteFileUseCase deleteFileUseCase;
 
     @Override
     @Transactional
     public MessageDto deleteRoom(Long id) {
+        Room room = findRoomPort.load(id)
+                .orElseThrow(() -> new AdminException(AdminErrorCode.ROOM_NOT_FOUND));
+
+        if (findItemPort.existByRoomId(id))
+            throw new AdminException(AdminErrorCode.HOLDING_ITEM_ROOM_DELETE);
+
         deleteRoomPort.deleteRoom(id);
+        deleteFileUseCase.deleteFile(room.getImageFile().getFileId().getId());
+
         return new MessageDto(id + "번 룸이 삭제되었습니다.");
     }
